[#1042242] Adding ability to Define words and letters to ignore
Added a form in Configure->Views Natural Sort->Configure Word Removal Lists
Allowed the ability for words specifically at the beginning of a phrase to be removed
Allowed the ability for words found anyehwere but the beginning of the phrase to be removed
Allowed for any symbol to be removed.
This all works globally.